Checks to see if link exists.

+++ scripts/postinstall.sh
@@ -4,4 +4,6 @@ cd /var/praekelt/goddard-captive-portal
 npm install
 
 mkdir -p /var/goddard/apps
-ln -s /var/praekelt/goddard-captive-portal /var/goddard/apps/captiveportal-redesign
+if [ ! -L /var/goddard/apps/captiveportal-redesign ]; then
+  ln -s /var/praekelt/goddard-captive-portal /var/goddard/apps/captiveportal-redesign
+fi
